QByteArray: fix count(ch) for more than 2Gi occurrences
We truncated the qsizetype result of countCharHelper() to int.

Fix by removing the int cast.

Fixes: QTBUG-105027
Pick-to: 6.4 6.3 6.2

+++ src/corelib/text/qbytearray.cpp
@@ -2625,7 +2625,7 @@ qsizetype QtPrivate::count(QByteArrayView haystack, QByteArrayView needle) noexc
 
 qsizetype QByteArray::count(char ch) const
 {
-    return static_cast<int>(countCharHelper(*this, ch));
+    return countCharHelper(*this, ch);
 }
 
 #if QT_DEPRECATED_SINCE(6, 4)
